safari-600.7-branch/LayoutTests/js/dfg-activation-register-overwritten-in-throw-expected.txt

This tests that the DFG does not attempt to overwrite the activation register with undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS try { g() } catch (e) { } is void 0 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
